Inital Commit Read Below
So this initial commit simply adds what I've been working on. Mainly trying to add to quality as some colors are extremely off. As it sits this basically has two functions, the Pearson Coefficient Correlation, and the Standard Error of Estimate. Both can be ran on here, but there are bugs and a coloring issue on tables.

The only bug I am currently trying to fix but will move on from to add more stuff is:

When graphing you can click data points, and it will show you the coordinates of these points. Now, when you change your excel sheet or change to different datas in the columns you'll see that you can no longer view the datapoints coordinates(x,y).

This should be fixed, the layout is also subject to change. Think of this as not even a prototype, like a protoprotoprototype.